+++ ricaun.NUnit/Models/TestModel.cs
@@ -53,7 +53,7 @@
         {
             var result = Skipped ? "Skipped" : Success ? "Passed" : "Failed";
             var name = string.IsNullOrEmpty(Alias) ? Name : Alias != Name ? $"{Alias}[{Name}]" : Alias;
-            var message = !Message.Contains("Exception") ? Message : "Exception";
+            var message = !Message.Contains("Exception") ? Message : Message.Substring(0, Message.IndexOf("Exception") + "Exception".Length);
             return $"{name}\t{result}\t{message}";
         }
     }