+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.2] / 2023-03-14
+### Features
+- [x] Enable Task Tests to work with the Framework
+- [x] Async Test Timeout - `TestEngineFilter.CancellationTokenTimeOut`
+### Updated
+- Update `TestModel`
+- Update `AppDomainUtils`
+- Add `TaskTests`
+- Add `InvokeTask`
+- Add `IsReturnTypeEqualsTask`
+- Add `IsParameterTypeSimilar`
+
 ## [1.2.1] / 2023-02-15
 ### Features
 - `AssemblyResolveService` ignore assembly version
@@ -133,6 +145,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Test with Console Output
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.2.2]: ../../compare/1.2.1...1.2.2
 [1.2.1]: ../../compare/1.2.0...1.2.1
 [1.2.0]: ../../compare/1.1.2...1.2.0
 [1.1.2]: ../../compare/1.1.1...1.1.2